backend/vector-search-service/requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
pydantic
